:art: refactor : InputView 기능 분리
