+++ src/test/java/ladder/step2/utils/InputUtilsTest.java
@@ -1,4 +1,4 @@
-package ladder.step2.view.input;
+package ladder.step2.utils;
 
 import ladder.step2.domain.Players;
 import ladder.step2.domain.PlayersTest;
@@ -11,13 +11,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-class InputViewTest {
+class InputUtilsTest {
     private static final String INPUT_EXCEPTION_MESSAGE = "올바른 입력 형식이 아닙니다. 다시 입력해주세요.";
     
     @Test
     @DisplayName("입력한 플레이어들의 객체를 반환받는다.")
     void inputPlayerNames() {
-        Players players = InputView.inputPlayerNames("pobi, honux, jun,jk");
+        Players players = InputUtils.getPlayers("pobi, honux, jun,jk");
         assertThat(players).isEqualTo(PlayersTest.PLAYERS);
     }
     
@@ -25,7 +25,7 @@ class InputViewTest {
     @DisplayName("플레이어 이름 입력 시, 구분자가 쉼표가 아니면 예외 던지기")
     void inputPlayerNamesDelimiterException() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputPlayerNames("pobi, jun, honux. jk"))
+                .isThrownBy(() -> InputUtils.getPlayers("pobi, jun, honux. jk"))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -33,7 +33,7 @@ class InputViewTest {
     @DisplayName("플레이어 이름 입력 시, 한 플레이어 이름의 글자 5자 초과 시 예외 던지기")
     void inputPlayerNamesLengthException() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputPlayerNames("pobi, tjdtls, honux, jk"))
+                .isThrownBy(() -> InputUtils.getPlayers("pobi, tjdtls, honux, jk"))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -41,7 +41,7 @@ class InputViewTest {
     @DisplayName("플레이어 이름 입력 시, 알파벳이 아닌 경우 예외 던지기")
     void inputPlayerNamesNonAlphabeticException() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputPlayerNames("pobi, jun, ho2ux, jk"))
+                .isThrownBy(() -> InputUtils.getPlayers("pobi, jun, ho2ux, jk"))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -50,14 +50,14 @@ class InputViewTest {
     @NullAndEmptySource
     void inputPlayerNamesNullOrEmptyException(String input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputPlayerNames(input))
+                .isThrownBy(() -> InputUtils.getPlayers(input))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
     @Test
     @DisplayName("최대 사다리의 높이 입력 값을 반환한다.")
     void inputLadderHeight() {
-        int ladderHeight = InputView.inputLadderHeight("11");
+        int ladderHeight = InputUtils.getLadderHeight("11");
         assertThat(ladderHeight).isEqualTo(11);
     }
     
@@ -65,7 +65,7 @@ class InputViewTest {
     @DisplayName("최대 사다리 높이 입력 시, 숫자가 아닌 값 입력 시 예외 던지기")
     void inputLadderHeightNonNumberException() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputLadderHeight("abc"))
+                .isThrownBy(() -> InputUtils.getLadderHeight("abc"))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -74,7 +74,7 @@ class InputViewTest {
     @NullAndEmptySource
     void inputLadderHeightNullOrEmptyException(String input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputLadderHeight(input))
+                .isThrownBy(() -> InputUtils.getLadderHeight(input))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -83,7 +83,7 @@ class InputViewTest {
     @ValueSource(strings = {"0", "04", "011"})
     void inputLadderHeightZeroException(String input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputLadderHeight(input))
+                .isThrownBy(() -> InputUtils.getLadderHeight(input))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
     
@@ -91,7 +91,7 @@ class InputViewTest {
     @DisplayName("최대 사다리 높이 입력 시, 음수를 입력 시 예외 던지기")
     void inputLadderHeightNegativeNumberException() {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> InputView.inputLadderHeight("-2"))
+                .isThrownBy(() -> InputUtils.getLadderHeight("-2"))
                 .withMessage(INPUT_EXCEPTION_MESSAGE);
     }
 }
